fixed Itay bug
get_x_cov returns a vector if a single covariate is used or a matrix if more than one are used.  The resampling that happens in get boot deg expects a matrix or data frame and thus forcing the cov output to be a data frame fixed the dimension error.